[services] Backup - Send attachments from pull backup reactor
Summary:
Depends on D4245

We want to send attachment holders for both logs and compactions.

The data is sent in the following order:
- compaction data
- compaction holders
- for log in logs
  - log data
  - log chunk

Test Plan: Services will still build properly. This change is eventually tested on the end of this stack when the logic for the backup test is added

Reviewers: palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, yayabosh

Differential Revision: https://phab.comm.dev/D4246