[docs] Update Flipper plugins instructions to navigate to Plugin Manager
Summary:
On the dev environment docs for Flipper plugins, it erroneously states:
{F73775}
There is no Manage Plugins in the View menu.
{F73776}
Instead, the Plugin Manager can be found here:
{F73777}
Updated the docs to reflect this change.

Linear issue with full context [[ https://linear.app/comm/issue/ENG-1260/dev-environment-docs-for-flipper-plugins-has-incorrect-steps-to-view | here ]].

Test Plan: N/A

Reviewers: ashoat, varun

Reviewed By: ashoat, varun

Subscribers: palys-swm, Adrian, atul

Differential Revision: https://phab.comm.dev/D4256