[yarn] Update `Podfile.lock` to reflect changes in D4244
Summary:
In D4244, I modified the `Podfile` in a way that didn't affect the dependencies, but the comments in the file. Because of this, the checksum at the bottom of `Podfile.lock` was modified (since the `Podfile` was changed). This diff includes that changed after running `yarn cleaninstall`.

Discussed lock files with @atul and understand that in the future I should run `yarn cleaninstall` when modifying any dependency management file.

Test Plan: Ran `yarn cleaninstall` in the home directory and got the updated checksum.

Reviewers: atul

Reviewed By: atul

Subscribers: ashoat, palys-swm, Adrian, atul

Differential Revision: https://phab.comm.dev/D4260