[CI] Bump GitHub Actions Ubuntu machines to `ubuntu-latest`
Summary:
We were previously setting this to `18.04`, but `20.04` is the latest stable release of Ubuntu. I figure we can just go with `ubuntu-latest` (which is currently an alias for `20.04`) to avoid having to specify a version number. Open to specifying `20.04` instead.. or any other possible approaches.

(Addresses note from @ashoat at: https://phab.comm.dev/D4239?id=13406#inline-25982)

Test Plan: Will land this and see what happens. **Could** fork the repo and populate the "Secrets" and whatnot to test this beforehand, but this diff is easy enough to revert if something goes wrong.

Reviewers: ashoat, palys-swm

Reviewed By: ashoat

Subscribers: Adrian, yayabosh, ashoat

Differential Revision: https://phab.comm.dev/D4261