[web] Fix sidebar title cutoff in `ChatThreadList`
Summary:
Linear issue with full context [[ https://linear.app/comm/issue/ENG-1270/sidebar-title-cut-off-from-chatthreadlist-on-web | here ]].

Seems to be a regression, not sure when it was introduced exactly. This diff repositions the `ChatThreadListItemMenu` `.menuSidebar` `div` to be absolutely positioned so that it doesn't interfere with the `ChatThreadListSidebar`. Before, the `ChatThreadListSidebar` was being covered by the `ChatThreadListItemMenu` and it probably went unnoticed since it wasn't tested with a sidebar message that was long enough to trigger the ellipsis overflow.

Before:
{F76328}

After:
{F76329}

Test Plan: Looks as expected on Chrome/Safari and the `ChatThreadListItemMenu` still works as expected.

Reviewers: atul, ashoat, palys-swm

Reviewed By: ashoat

Subscribers: palys-swm, Adrian

Differential Revision: https://phab.comm.dev/D4284