[web] Add `LoadingIndicator` to `ThreadSettingsGeneralTab` Save button
Summary:
Related Linear issue [[ https://linear.app/comm/issue/ENG-1271/add-loadingindicator-to-threadsettingsgeneraltab-save-button | here ]]. Adds a `LoadingIndicator` to the Save button in `ThreadSettingsGeneralTab` to indicate that a change is in progress. Depends on D4105 because of the renaming of `inputDisabled` to `threadSettingsOperationInProgress`.

Before:
{F76421}

After:
{F76422}

Test Plan: Tested on Chrome/Safari and works as expected. However, since `localhost/comm` is faster than `prod`, the change can be seen less quickly in the videos. However, this diff is very similar to D3605 which works fine on `prod`.

Reviewers: atul, ashoat

Reviewed By: atul

Subscribers: palys-swm, Adrian

Differential Revision: https://phab.comm.dev/D4287