[services] Change services yarn script name from run-local-cloud to init-local-cloud
Summary:
When running yarn run-local-cloud we are starting the localstack docker image and initialize the DynamoDB tables and S3 buckets using the Terraform
scripts.

When we are starting some of the services the localstack image starts automatically along with it because for each service we have in docker-compose
depends on the parameter:

```
depends_on:
      - localstack
```

That's why we need to run yarn run-local-cloud only for the first-time initialization.

For the tunnelbroker we have depends_on for a RabbitMQ local docker image as well (but it doesn't need to be initialized, just started along with the
service).

From the described above the name of the command doesn't reflect what it is doing and the name changed to init-local-cloud in the services package.json
would be more purpose-reflecting.

Linear task: [[ https://linear.app/comm/issue/ENG-1275/change-services-yarn-script-name-from-run-local-cloud-to-init-local | ENG-1275 ]]

Test Plan: Run `yarn init-local-cloud` and the expected result is localstack docker image start and dynamoDB tables, s3 buckets are initialized by the Terraform script.

Reviewers: karol-bisztyga, palys-swm, varun

Reviewed By: karol-bisztyga, palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, yayabosh

Differential Revision: https://phab.comm.dev/D4303