[web] Handle relationship update errors
Summary:
We should close the modal only after successful update. We can consider keeping the modal open and closing the tags, but we definitely must not close the modal before the response is received.

Depends on D4314

Test Plan: Modify relationship responder so that every second request succeeds. Select a user and click confirm button - the modal should remain open and the button should be disabled until the response is received. After the response, click the button again - the modal should get closed and the user should appear on friends list.

Reviewers: def-au1t, atul, ashoat

Reviewed By: atul, ashoat

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phab.comm.dev/D4315