[lib] Update `messageStore.messages` with `msgStoreOps` for `sendText/MultimediaMessage.started` actions
Summary: Broke this into its own diff since it's a little more involved. Basically continuing refactoring work to get to the point where we can remove `assertMessageStoreMessagesAreEqual` and solely use the "ops" to update `messageStore.messages` in redux.

Test Plan: This should be a straightforward refactor given we've had `assertMessageStoreMessagesAreEqual` in place for some time now.

Reviewers: def-au1t, palys-swm, yayabosh

Reviewed By: palys-swm

Subscribers: ashoat, Adrian

Differential Revision: https://phab.comm.dev/D4330