[web] [landing] Omit units for 0 values in CSS
Summary:
The units for 0 values in CSS are redundant. This is a quick no-op change that makes the code more concise and readable as the unitless 0 is distinguished from hard-set values with units.

Read more: https://stackoverflow.com/q/7923042/13079675

---

This technically depends on D4339 since both diffs modify `landing/subscription-form.css`, but I can resolve that merge conflict quickly and easily.

Test Plan: Find-and-replace + close reading. `0px == 0` in CSS, so there are no style changes that can occur as a result of this.

Reviewers: atul, palys-swm

Reviewed By: palys-swm

Subscribers: ashoat, Adrian

Differential Revision: https://phab.comm.dev/D4343