[web] Remove `safari-pinned-tab.svg`
Summary:
Addresses https://linear.app/comm/issue/ENG-963/remove-old-squad-cal-favicon-on-safari

It used to be that Safari didn't display favicons for each tab and instead showed the page title. However, if a tab was pinned--or there wasn't enough space to show an abbreviated title--it would display the first letter of the page title:

{F83038}

{F83039}

Apple gave the option of providing a black square SVG (`safari-pinned-tab.svg`) to be displayed for "pinned" tabs and the ability to specify a tint color for "active" tabs.

However, recent versions of Safari display favicons for tabs across the board (it's no longer even possible to disable them), and including `safari-pinned-tab.svg` is no longer necessary... but Safari will still use it if it's provided. Based on what I've read, it seems like the only reason to specify `safari-pinned-tab.svg` is if you want something other than the default favicon to be displayed on just Safari. I don't think that's something we care about.

Test Plan: Ensured correct favicon displayed in Safari/Chrome after clearing cache.

Reviewers: yayabosh, palys-swm, def-au1t

Reviewed By: yayabosh, palys-swm

Subscribers: ashoat, Adrian

Differential Revision: https://phab.comm.dev/D4358