[services][identity] update Config struct to hold keypair instead of just private key
Summary: We need to access both the server public and private key to do PAKE registration and login. Although it's possible to derive the public key from the private key, the opaque-ke library only lets us derive the full keypair from a byte slice. We should do this once when we load the config from a file at server start time.

Test Plan: ran server and confirmed the keypair was loaded successfully

Reviewers: jimpo, palys-swm, karol-bisztyga

Reviewed By: palys-swm

Subscribers: ashoat, Adrian, atul, yayabosh

Differential Revision: https://phab.comm.dev/D4362