Implement elementary file operations with encryption-decryption layer
Summary: Implements file operations that work with encryption-decryotion layer defined in previous diffs

Test Plan: Modify NotificationService and AppDelegate so that the first writes to the file upon notification and the second read it on AppLaunch. Examine in the debugger that the content read by AppDelegate is consistent sent notifications.

Reviewers: palys-swm, varun, karol-bisztyga

Reviewed By: palys-swm

Subscribers: ashoat, Adrian, atul, yayabosh

Differential Revision: https://phab.comm.dev/D4385