[services] Tunnelbroker - Add test for putMessageItemsByBatch function
Summary:
This is a test for `putMessageItemsByBatch` function introduced in D4378.
In this test, we are creating a messages batch of 29 items and putting them into the database by the `putMessageItemsByBatch` function. We are expecting the
size of the found items would be the same as an item's batch size.

Linear task: [[ https://linear.app/comm/issue/ENG-1301/put-tunnelbroker-messages-in-a-batch-to-dynamodb-database | ENG-1301 ]]

Test Plan: Run `yarn test-tunnelbroker-service-dev-mode`

Reviewers: karol, tomek

Reviewed By: tomek

Subscribers: ashoat, tomek, adrian, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D4377