[services] Add dev-mode to the services documentation
Summary:
This diff introduces the services documentation section for the dev-mode for services.
Describe how to use the localstack and rabbitMQ cloud for development and testing purposes.

Linear task: [[ https://linear.app/comm/issue/ENG-823/update-services-documentation-with-the-dev-mode | ENG-823 ]]

Test Plan: Running commands in the documentation work as expected according to the doc.

Reviewers: karol-bisztyga, palys-swm, ashoat, varun

Reviewed By: palys-swm, ashoat, varun

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, yayabosh

Differential Revision: https://phab.comm.dev/D4253