[services] Backup - Pull Backup - Wait for the get reactor to terminate
Summary:
Depends on D4326

This is a bug fix which was spotted in https://phab.comm.dev/D4228#121209

The problem was, we didn't wait for the `getReactor` to gracefully terminate in the `PullBackupReactor`. The result was, if the client stopped working, the whole service crashed. Now it just shows an error and all objects are deleted properly.

Test Plan: Start pulling the backup and crash the client (you can probably use ctrl+c but I just used `panic` in rust) in the middle of receiving compaction chunks.

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, yayabosh

Differential Revision: https://phab.comm.dev/D4341