[keyserver] Run `source-nvm.sh` using `bash` in `Docker` container
Summary:
Relevant Linear issue with full context [[ https://linear.app/comm/issue/ENG-1339/run-source-nvmsh-using-bash-in-docker-container | here ]]. This diff modifies the `Dockerfile` in `keyserver` to execute `source-nvm.sh` in the environment specified by the shebang in `source-nvm.sh` instead of in the current environment of the container, which uses `sh`.

Read more [[ https://stackoverflow.com/questions/54109341/bash-shebang-is-ignored-script-is-still-executed-with-zsh/62244534#62244534 | here ]].

Thank you @atul for your help on this!

Test Plan:
Ran the Docker keyserver build locally and it worked with no errors.
{F89800}

Reviewers: atul, ashoat

Reviewed By: ashoat

Subscribers: palys-swm, Adrian

Differential Revision: https://phab.comm.dev/D4429