[services] Tunnelbroker - Add exponential backoff related constants
Summary:
This diff introduces additional constants for the function for exponential backoff timeout which will be used in DynamoDB batch write operation and throttling. The function itself is introduced in D4373, and the usage of these constants is in D4378.

`DYNAMODB_BACKOFF_FIRST_RETRY_DELAY` - Is an exponential backoff first delay in milliseconds.
`DYNAMODB_MAX_BACKOFF_TIME` - Maximum delay timeout after which the operation would be considered as failed.

Test Plan: No testing is needed, it's just a constant.

Reviewers: karol-bisztyga, palys-swm

Reviewed By: karol-bisztyga, palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, yayabosh

Differential Revision: https://phab.comm.dev/D4436