[services] Tunnelbroker - Switching to use composite key in a messages table in Terraform
Summary:
To avoid possible message UUID collision and speedup batch messages deletion we should switch to using the composite key for the `tunnelbroker-messages` table which will consist of `messagesID` as a Sort Key and `toDeviceID` as a Partition key.

This diff is a part of the stack and introduces changes for Terraform.

Linear task: [[ https://linear.app/comm/issue/ENG-1362/update-to-use-composite-key-messageidtodeviceid-for-messages-table-in | ENG-1362 ]]

Test Plan:
These changes should be tested at the D4475 diff.  The current `messageItem` related tests at `services/tunnelbroker/test/DatabaseManagerTest.cpp` should not fail including:
- PutAndFoundMessageItemsStaticDataIsSame
- PutAndFoundMessageItemsGeneratedDataIsSame
- PutAndFoundByReceiverMessageItemsDataIsSame

Reviewers: karol, tomek

Reviewed By: karol, tomek

Subscribers: ashoat, tomek, adrian, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D4473