[web] Reintroduce `MessageTimestampTooltip` in `ChatMessageList`
Summary:
This diff is the first part in a stack of commits that will work to fix the timestamp bug outlined in this [[ https://linear.app/comm/issue/ENG-1266/timestamp-appears-below-chatinputbar-on-web | Linear issue ]]. Also see @ashoat's [[ https://phab.comm.dev/D3567#125491 | comment ]] on D3567 for more context.

This diff removes the timestamp from `MessageTooltip` and instead reintroduces `MessageTimestampTooltip` in `ChatMessageList`, thus reverting the visual aspect of this change to what the timestamp tooltip used to look like before `MessageActionButtons` was introduced.

The reason this was done is:
1. Start the new stack of diffs to fix the timestamp bug
2. By moving the timestamp out of `MessageTooltip`, the component is no longer inside the `messageContainer` in `ChatMessageList` which has a `overflow-y: auto` property. This `overflow-y` property was causing the timestamp to be cut off within the `ChatMessageList` container since it did not allow any elements to extend beyond its perimeter.
3. This reverts the code to a familiar place that we have seen before. This diff should be fairly easy to review -- although there are added files, these are mostly just copied over from the old version of the tooltip/timestamp and the front-end changes are identical.

Test Plan:
Tested on Chrome/Safari and looks as expected.
{F95071}

Reviewers: tomek, jacek, atul

Reviewed By: tomek, atul

Subscribers: atul, adrian, ashoat

Differential Revision: https://phab.comm.dev/D4481