Remove unnecessary (dead code) code from message reducer.
Summary: This differential removes code that was used to modify messageStore on rehydrateActionType. Now that we do not persist messages in messageStore in redux, messages are always {} on rehydrateActionType, so the code never get executed.

Test Plan: This differential does not change app functionality, but only the way it works. So we expect the app to work exactly the same after this change.

Reviewers: tomek, atul, jacek, ashoat

Reviewed By: atul, ashoat

Subscribers: ashoat, adrian, abosh

Differential Revision: https://phab.comm.dev/D4507