Filter local multimedia messages
Summary:
This diff places the code responsible for local multimedia messages to setMessageStoreMessages action handling, since it is the only place we can do so, that we store messages in SQLite now.

---

Depends on D4508

Test Plan: After this differential app is also expected to work in the same manner as before.

Reviewers: atul, tomek, jacek

Reviewed By: atul

Subscribers: ashoat, adrian, abosh

Differential Revision: https://phab.comm.dev/D4508