[keyserver] Add dbType to DBConfig
Summary:
While working on this stack I've realized there are some incompatible changes we'll need to make (where MariaDB needs one query and MySQL needs another).

To make the transition smooth for devs, and to maintain a potential path to migrate prod back if we see a serious issue after deploying, I'm going to make sure this stack maintains support for both MySQL and MariaDB in the near term.

So the Node application will need to have some awareness of MySQL 5.7 vs. MariaDB 10.8.

Depends on D4515

Test Plan: I added a `console.log` of `dbConfig` in `loadPool` and made sure it included `"dbType":"mariadb10.8"`

Reviewers: jonringer-comm, atul, palys-swm

Reviewed By: atul

Subscribers: Adrian, yayabosh

Differential Revision: https://phab.comm.dev/D4516