[keyserver] Avoid JSON cast when using MariaDB
Summary:
One difference between MySQL and MariaDB is that MySQL has a `JSON` column type and MariaDB doesn't.

MySQL requires that the second parameter to `JSON_CONTAINS` is a `JSON` type. MariaDB doesn't have `JSON` types, so it doesn't have this requirement.

We need to make sure we do the cast for MySQL, but not for MariaDB.

Depends on D4521

Test Plan: I ran the web application and clicked around a bunch. Full test plan will come later

Reviewers: jonringer-comm, atul, palys-swm

Reviewed By: atul

Subscribers: Adrian, yayabosh

Differential Revision: https://phab.comm.dev/D4522