[services] Tunnelbroker - Add Cargo manifest file
Summary:
This diff introduces a Rust Cargo manifest file for the notification library wrapper in Rust for using it from C++ using corrosion tool.

Linear task: [[ https://linear.app/comm/issue/ENG-1330/rust-notification-library-integration | ENG-1330 ]]

Test Plan: No tests, this is a manifest-only file. This file will be updated with the `dependencies` and built using `cargo build` in a follow-up diff of the stack.

Reviewers: karol-bisztyga, varun, jonringer-comm

Reviewed By: varun, jonringer-comm

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, yayabosh

Differential Revision: https://phab.comm.dev/D4534