[chore] Replace `'` with `’` in user-displayed text
Summary: Replaced single quote with apostrophe in user-displayed text. Wasn't sure the best way to title this, but basically this diff only replaces the single quotes in text that can get displayed in a modal, the app, or an error message that //isn't// an `invariant`.

Test Plan:
Used this regex to search for all instances of `'` surrounded by `""`:
{F102843}

Reviewers: atul

Reviewed By: atul

Subscribers: ashoat, palys-swm, Adrian

Differential Revision: https://phab.comm.dev/D4541