overwrite background notification priority anbd pushType properties so that they match apple docs
Summary: This differential updates headers of APNs notification push request so that they are consitent with apple docs regarding background notifications. We set notification priority to 5, so that decision upon its delivery is dependent on power consumption and push-type to 'background' so that we do not risk notification being discarded by the APNs.

Test Plan: This differential does not introduce new functionality. We can open two clients - web wnd mobile - to ensure correct rescinding mechansim.

Reviewers: tomek, ashoat

Reviewed By: ashoat

Subscribers: ashoat, adrian, atul, abosh

Differential Revision: https://phab.comm.dev/D4557