[keyserver] Create a new session if the old one is expired
Summary:
See [Linear issue](https://linear.app/comm/issue/ENG-1408/viewersessioninfo-should-be-set). I was seeing a bug where if a web session was too old (older than 3 days) and the session got deleted, the client would be unable to reconnect.

This diff fixes the issue by making the server create a new session if the old one is expired.

Test Plan: I had a client that was reproing the issue (old tab on an old laptop). I patched this fix onto my keyserver and watched the client magically connect and update

Reviewers: atul, palys-swm, def-au1t

Reviewed By: atul

Subscribers: Adrian, yayabosh

Differential Revision: https://phab.comm.dev/D4570