[services] Backup - Fix terminate callback in pull backup reactor
Summary: We have to check whether the get reactor is initialized in the pull backup reactor before checking its status. The initialization of this reactor happens in the `writeResponse` method so if we have an error in the `initialize` method, `terminate`'s going to be called with get reactor uninitialized and that will lead to a crash.

Test Plan: make a request for pulling a backup without passing a backup id (missing any fields would trigger this I think).

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, yayabosh

Differential Revision: https://phab.comm.dev/D4223