[lib] Add comments in messageTypes definition
Summary:
Add comments in messageTypes definition to clarify the difference between CREATE_SUB_THREAD and
CREATE_THREAD/CREATE_SIDEBAR. The CREATE_SUB_THREAD is now referencing only creating subthreads, but it used to also reference
creating sidebars. This can be observed in the database. For this reason this usecase has to be acknowledged in the code (eg.
in lib/shared/messages/create-sub-thread-message-spec.js).

Test Plan:
Check if the types actually correspond to specified chat types by changing the robotext displayed to the user based
on them.

Reviewers: ashoat, palys-swm

Reviewed By: ashoat

Subscribers: Adrian, atul, yayabosh

Differential Revision: https://phab.comm.dev/D4575