[lib, native, web] Remove showingSidebarsInline variable and button text option "See threads..."
Summary:
There was some obsolete code that got never called since diff D1487. It got removed.
showingSidebarsInline was indeed obsolete as the only function that used it - ChatThreadListSeeMoreSidebars - was called inside an ‘if’ not allowing sidebarsItems.length to be 0 and the variable was being set as showingSidebarsInline: sidebarItems.length !== 0
Thus the code defined for when showingSidebarsInline == 0 got removed as well.

Linear issue: https://linear.app/comm/issue/ENG-1426/delete-see-threads-button-text-option

Test Plan: Check that everything still compiles.

Reviewers: ashoat, palys-swm

Reviewed By: ashoat

Subscribers: Adrian, atul, yayabosh

Differential Revision: https://phab.comm.dev/D4577