[services] Tests - Add GET for Blob performance tests
Summary:
Depends on D4582

Adding the logic for the `get` operations.

Test Plan:
```
cd services
yarn run-performance-tests blob
```

You may run this but running it again will result in a failed test. This is because the data are not going to be removed in the end so you'll get "the data already exists" error. In this case, you may want to run:
```
yarn reset-local-cloud
```

Reviewers: tomek, varun

Reviewed By: tomek

Subscribers: ashoat, tomek, adrian, atul, abosh

Differential Revision: https://phab.comm.dev/D4583