[services] Build arm containers for m1 mac
Summary: The point is to not launch intel-based containers on m1 machines. I rebuilt the base image on m1 and pushed it to the hub with a different tag. Doing a simple check in bash scripts we can identify whether we're on m1 or not and use proper version of base image.

Test Plan:
You have to be on the m1 mac.
Discard the changes from this diff and do this:
```
cd services
yarn run-backup-service
```
Go to the docker desktop and see that the container has a `amd` label. When hovered, it'll say that this image may be inefficient due to virtualization.

Remove this container, apply the changes from this diff and repeat all operations. There's not going to be that label for the new container. I also observed a significant improvement in container's performance.

Reviewers: geekbrother, ashoat

Reviewed By: geekbrother, ashoat

Subscribers: jimpo, jonringer-comm, ashoat, palys-swm, Adrian, atul, yayabosh

Differential Revision: https://phab.comm.dev/D4375