Introduce iOS and Android notification max size thresholds as constants
Summary: According to documentations, iOS max notification size is 4096 bytes, while for Android (FCM) it is 4000 bytes. Including this constants in 'utils' since there are some other constants related to notifications.

Test Plan: There is no test plan for this issue. Links to documentation that support values are in the Linear

Reviewers: tomek

Reviewed By: tomek

Subscribers: ashoat, adrian, atul, abosh

Differential Revision: https://phab.comm.dev/D4600