[web] add redux persist dependency
Summary:
We would like to use persistence on web and the most convenient way to do that is to use redux persist. We already use that library on native. The default persistence location of web is local storage which is what we need.

Test Plan: Run the web and check if opening the app works. The behavior shouldn't be affected by this change.

Reviewers: jacek, atul, ashoat