[services] add GetUserID RPC
Summary:
Users and keyservers need to be able to retrieve the userID corresponding to some given user info (username in the case of password auth and wallet address in the case of wallet auth). This RPC exposes that functionality.

Depends on D4613

Test Plan: Checked syntax with linter, identity service build succeeded

Reviewers: ashoat, tomek, karol, max, derek

Reviewed By: ashoat, tomek

Subscribers: adrian, atul, abosh

Differential Revision: https://phab.comm.dev/D4614