Add a function for creating unique device ID on mobile in cpp/jsi
Summary:
Add a function for creating unique device ID on mobile in cpp/jsi
Linear issue: https://linear.app/comm/issue/ENG-1279/initialize-deviceid-on-clients

Test Plan:
Log to the console the functions output given 'KEYSERVER', 'WEB', 'MOBILE' and some other random
string. See that in the first three cases it produces an ID compatible with what was described in
the linear issue comment, and that in the third case the app throws an error - as we cannot proceed
without a device ID (assuming the tunnelbroker is being used when this function gets called).

Reviewers: tomek, max

Subscribers: ashoat, adrian, atul, abosh

Differential Revision: https://phab.comm.dev/D4616