[web] Move selected chat to the top of the list in creation mode
Summary:
When selecting user for new chat, an existing chat can be matched and selected. However it is often not visible in chat list because in's not near the top.
The diff fixes this by moving active thread in chat creation to the top of the chat list
This is the follow-up diff after: https://phab.comm.dev/D4461#131918

Test Plan: Start chat creation, select set of users matching existing chat, that is not near the top of the list. It should be at the top after this diff.

Reviewers: tomek, atul

Reviewed By: tomek

Subscribers: ashoat, adrian, abosh

Differential Revision: https://phab.comm.dev/D4615