[keyserver] Use mysqldump from MariaDB 10.8 instead of MySQL 8
Summary:
In D4183 we decided to use `mysqldump` from MySQL 8 instead of MySQL 5.7 because it was too challenging to get the right one.

Now that we're using MariaDB 10.8, it's easier to get the right `mysqldump` for the given database server.

Test Plan: I tested a backup from this updated image and then tested a restore of that backup on the same image

Reviewers: atul, tomek

Reviewed By: tomek

Subscribers: adrian, abosh

Differential Revision: https://phab.comm.dev/D4641