[Nix] Update nixpkgs pin
Summary:
Update the dev shell.

This also includes the changes from
https://github.com/NixOS/nixpkgs/pull/182364 which fixes arcanist.

Arcanist is passing on official nix cache:
https://hydra.nixos.org/eval/1773081?filter=arcanist&compare=1773037&full=#tabs-still-succeed
so it's available for all platforms

The pruned node patch is only relevant for environments which are missing
`$HOME` (e.g. nix's sandbox builds), but shouldn't affect dev workflow. Also,
may not be relevant, as patch fails to apply as it's "already applied".

Depends on D4660

Test Plan: `nix develop`

Reviewers: varun, ashoat

Subscribers: tomek, adrian, atul, abosh

Differential Revision: https://phab.comm.dev/D4646