[services] Tunnelbroker - Fix `IDs` in Terraform dynamodb table attributes
Summary:
This diff is a fix.
After updating the field names to use ID  instead of Id some of the fields used the old names which caused errors while running in the cloud. The codebase is already
using the ID  pattern. The changes need to be made in the Terraform script.

Linear task: [[ https://linear.app/comm/issue/ENG-1492/fix-ids-in-terraform-dynamodb-tunnelbroker-table-attributes-to-use-id | ENG-1492 ]]

Test Plan: Run `yarn init-local-cloud` and run database related unit tests in `test/DatabaseManagerTest.cpp`.

Reviewers: karol, tomek

Reviewed By: tomek

Subscribers: ashoat, tomek, adrian, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D4693