[Nix] Re-package arcanist as just the php scripts
Summary:
Nixpkgs `arcanist` pacakge includes "phage" shell which
is a fairly large c++ and php code base. However,
we only really care about using the `arc` command, so we should reduce
the scope of the `arcanist` package to be what we actually use.

This avoids situations such as https://phab.comm.dev/D4483#131136
where the build for phage (and it's requirements), aren't thread safe. They
use bison and flex, and the generated files aren't delcared in a way in which
make will be aware that generation of those files need to be ordered before
usage of them.

Test Plan:
```
# this can be used from any branch
nix develop github:commE2E/comm?ref=update-nix-flakes
# use any arc command, 8.1 would fail on `arc diff`
```

Reviewers: varun, ashoat

Reviewed By: ashoat

Subscribers: tomek, adrian, atul, abosh

Differential Revision: https://phab.comm.dev/D4662