Disable shared preferences data from being automatically backed up in google cloud
Summary: From Android API 23 and so on data that application writes to shared preferences are backed up in google cloud if not deleted prior to application being uninstalled. We do not want this behaviour since we are about to deliver backup service for that purpose.

Test Plan: temporarily add a line that logs encryption key in SQLiteQueryExecutor::initialize(). Ensure in Android log cat that encryption key changes with every application re-install

Reviewers: tomek, atul

Subscribers: ashoat, adrian, abosh

Differential Revision: https://phab.comm.dev/D4663