[web] Remove creating a new siedebar/thread with an empty message
Summary:
It was possible on web to create a new sidebar or thread with an empty (or including only
whitespaces) message. It is no longer so.
Linear issue:
https://linear.app/comm/issue/ENG-1437/sending-empty-message-on-web-is-possible-for-pending-chat-and-creates

Test Plan:
Check that it is no longer possible to create a new sidebar/thread with an empty/including only whitespaces message, but is
possible with a non-empty text message or a multimedia message.
Additionally check that sending text and multimedia messages is possible, including those where
the multimedia is chosen from a finder, pasted from a clipboard and drag-and-dropped.

Reviewers: jacek, tomek

Reviewed By: jacek, tomek

Subscribers: ashoat, tomek, adrian, atul, abosh

Differential Revision: https://phab.comm.dev/D4640