[web] Fix: Inbox spacing regression
Summary:
Fix for issue: https://linear.app/comm/issue/ENG-1473/inbox-spacing-regression-on-safari-on-web
~~I used CSS child selector to avoid modifying reusable `Tabs` component~~
Modified `Tabs` component CSS by making it fill the container.

Test Plan:
Log in as a user that doesn't have any thread (or first message in the thread) with long name. The chat list should now fill the container.
Tested on Chrome, Safari and Firefox.

Reviewers: tomek, atul, ashoat

Reviewed By: tomek, ashoat

Subscribers: adrian, abosh

Differential Revision: https://phab.comm.dev/D4666