[web] Make "Mark as unread" button easier to click reliably
Summary:
Context: https://linear.app/comm/issue/ENG-992/mark-as-unread-button-can-be-hard-to-press-on-web

Users reported that sometimes the "Mark as unread" button would disappear before they were able to click it. It looks like this was because there was a `1px` gap (see image below) between the menu button (vertical ellipsis) and the toggle unread status button ("Mark as unread") where `onMouseLeave` would trigger and hide the toggle unread status button.

{F120109}

I was able to reproduce this issue consistently when moving the cursor slowly from the vertical ellipsis to "Mark as unread" button, but was able to avoid the issue by moving my cursor "fast." Not sure why `onMouseLeave` doesn't trigger if you move the cursor back within the `div` fast enough, but that's the behavior I observed.

This diff basically closes that `1px` gap and gets things working reliably.

Test Plan: Was able to consistently mark a thread as read/unread without the "Mark as unread" button disappearing.

Reviewers: tomek, abosh, varun, jacek, ashoat

Reviewed By: abosh, varun, ashoat

Subscribers: ashoat, adrian

Differential Revision: https://phab.comm.dev/D4669