[keyserver] Add index to container column of uploads table
Summary: This is the first part of the stack to resolve [ENG-1484](https://linear.app/comm/issue/ENG-1484/cant-scroll-up-in-chat-history). We'll start using it in `message-fetchers.js` in the following diff.

Test Plan: I first tested this stack locally in my environment to make sure the queries worked. Then I patched it onto my production keyserver and confirmed performance was improved

Reviewers: tomek, atul

Reviewed By: tomek

Subscribers: adrian, abosh

Differential Revision: https://phab.comm.dev/D4673