[keyserver] Look up uploads for a message using the container column
Summary:
This is the second part of the stack to resolve [ENG-1484](https://linear.app/comm/issue/ENG-1484/cant-scroll-up-in-chat-history), and is the core of it. This approach has ***way*** better performance.

Depends on D4673

Test Plan: I first tested this stack locally in my environment to make sure the queries worked. Then I patched it onto my production keyserver and confirmed performance was improved

Reviewers: tomek, atul

Reviewed By: tomek

Subscribers: adrian, abosh

Differential Revision: https://phab.comm.dev/D4674