[services] Tunnelbroker - Fix table names to use config-only instead of constant
Summary:
This diff is a fix.
Fix to use table names from config-only everywhere in the codebase instead of constant with the default table name. This will fix the bug when the default table name from constants is used instead of the table name for the test because in the sandbox mode Tunnelbroker reads the corresponding config file from `services/tunnelbroker/tunnelbroker-dev.ini`.

Linear task: [[ https://linear.app/comm/issue/ENG-1494/fix-to-use-table-names-from-config-only-everywhere-in-tunnelbroker | ENG-1494 ]]

Test Plan: Run tests in a sandbox using the `yarn run-unit-tests` and use the table names from the `services/tunnelbroker/tunnelbroker-dev.ini` config file.

Reviewers: karol, tomek

Reviewed By: tomek

Subscribers: ashoat, tomek, adrian, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D4696