[services] Tunnelbroker - AMQP message payload string conversion to use size and remove double `getPayload()` call
Summary:
This is a small improvement and possibly a bug fix. We should use a message body size when constructing a string from an AMQP message payload instead of using body-only.

As a small follow-up presented in this diff: fix `message->getPayload()` double call and use the local instance instead. It's a super small change that's why merged it into a single one but a small diff to save a reviewer's time.

Linear task: [[ https://linear.app/comm/issue/ENG-1495/amqp-cpp-client-throwing-a-segmentation-error-in-a-highmid-load-and | ENG-1495 ]]

Test Plan: Successfully built the server using `yarn run-tunnelbroker-service-in-sandbox` command. Run tests using `yarn run-unit-tests`

Reviewers: karol, tomek

Reviewed By: tomek

Subscribers: ashoat, tomek, adrian, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D4702