[services] Tunnelbroker - Fix message `from` and `to` swap in AMQP tests
Summary:
This diff is a fix.

The constructor for the `MessageItem` entity was changed (`fromDeviceID` and `toDeviceID was swapped in places), but the tests have not reflected these changes. This diff fixes tests for the message items to reflect these changes.

Related Linear task: [[ https://linear.app/comm/issue/ENG-1491/deploy-and-ci-testing-related-fixes | ENG-1491 ]]

Test Plan: Tests in `DatabaseManagerTest.cpp` was successfully built and passed in a sandbox.

Reviewers: karol, tomek

Reviewed By: karol, tomek

Subscribers: varun, ashoat, tomek, adrian, atul, karol, abosh

Differential Revision: https://phab.comm.dev/D4745