[web] Remove `connectDropTarget` from `ChatMessageList` with new `tooltip` changes
Summary: In D4434, there were certain changes made that made D4481 broken since `ChatMessageList` was refactored. This diff brings `ChatMessageList` up to date and out of a broken state. (This is also the reason D4697 and D4698 were broken.) This is the first diff in a stack of diffs that will resolve the timestamp bug, related Linear issue [[ https://linear.app/comm/issue/ENG-1266/timestamp-appears-below-chatinputbar-on-web | here ]].

Test Plan: Tested on `web` and everything looks as expected.

Reviewers: atul, jacek, tomek

Subscribers: