[CI] Add removal of irrelevant Harbormaster git tags to GH Actions
Summary:
On every push to `master` we'll remove Harbormaster tags (of the form `phabricator/base/123` and `phabricator/diff/123`) associated with diffs that are "Abandoned"/"Closed".

---

Depends on D4689

Test Plan:
This one is kind of hard to test. Usually test GH Actions workflows on a fork of the repo, but the forked repo didn't hold onto the phabricator tags from `CommE2E/comm`.

If there are issues, reverting this diff would be simple.

Reviewers: tomek, ashoat, jacek, abosh, varun

Reviewed By: varun

Subscribers: ashoat, adrian

Differential Revision: https://phab.comm.dev/D4690