[web] Finish styling `MessageTimestampTooltip`
Summary:
Relevant Linear issue [[ https://linear.app/comm/issue/ENG-1266/timestamp-appears-below-chatinputbar-on-web | here ]].

This diff adds styling changes to `MessageTimestampTooltip`, specifically changing the color of the tooltip, increasing the width of the tooltip (so it is at minimum the width of the `MessageTooltip` with the reply/create thread buttons), some padding, and makes the border radius match the `MessageTooltip`.

Depends on D4704.

Test Plan:
Tested on Chrome/Safari and looks as expected:
Chrome:

Safari:

Reviewers: atul, jacek, tomek

Subscribers: ashoat, adrian

Differential Revision: https://phab.comm.dev/D4705