[keyserver] Clean up `scp-backup.sh` using ShellCheck
Summary: Related Linear issue [[ https://linear.app/comm/issue/ENG-1295/clean-up-shell-scripts-in-codebase | here ]]. This is part of a set of diffs that will allow ShellCheck to be added to the CI. See inline comments for specific details of the ShellCheck error/warning output.

Test Plan:
[[ https://www.shellcheck.net/ | ShellCheck ]]

Additionally, this script is run during the keyserver build so if that workflow continues to pass CI we can reasonably assume that the script continues to work as expected.

Reviewers: atul, jon, tomek, ashoat

Reviewed By: atul, ashoat

Subscribers: adrian

Differential Revision: https://phab.comm.dev/D4713