[services] Clean up `run_performance_tests.sh` using ShellCheck
Summary: Related Linear issue [[ https://linear.app/comm/issue/ENG-1295/clean-up-shell-scripts-in-codebase | here ]]. This is part of a set of diffs that will allow ShellCheck to be added to the CI. See inline comments for specific details of the ShellCheck error/warning output.

Test Plan:
Similarly as in D4724, I went to @karol's diff D4580 where this code was introduced. Looks like @karol noted:

> the desired behavior for them is to fail with a message "unimplemented".

So I'm assuming this was supposed to break, and that is why it is still failing? I don't have much context on this. Either way, this diff is scoped to the ShellCheck side of the script, so if it's supposed to fail/be broken then I think that's outside the scope of this diff.

Reviewers: atul, jon, tomek, karol

Reviewed By: jon, karol

Subscribers: ashoat, adrian, karol

Differential Revision: https://phab.comm.dev/D4725