[web] Include `Relationship` tab in thread settings modal
Summary:
Display relationship tab in thread settings modal if there are some relationship actions available.
It modifies size of the modal so it has now the same size in every tab - as we discussed it already with Ashoat.

{F120833}

Test Plan: In personal threads additional tab "Relationship" should appear in thread settings modal. The actions should be exactly the same as in mobile app.

Reviewers: tomek, atul, ashoat

Reviewed By: tomek, ashoat

Subscribers: ashoat, adrian, abosh

Differential Revision: https://phab.comm.dev/D4681